requirements.txt
streamlit==1.35.0
gTTS==2.5.1
pillow==10.3.0